Move DatabaseAnalyzer, Add column object
